requirements.txt
requests
Pillow
matplotlib
numpy
scikit-image
tqdm
colorthief